new: Update favorites tab-navigation
