+++ src/index.js
@@ -16,7 +16,6 @@ const hosts = generateHosts(10);
 
 const store = createStore(reducer,
     window.__REDUX_DEVTOOLS_EXTENSION__ ? window.__REDUX_DEVTOOLS_EXTENSION__() : (f) => f);
-store.subscribe(() => console.log(store.getState()));
 
 ReactDOM.render(
     <Provider store={store}>
